+++ src/components/MARCTableView/MARCTableViewRow.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../test/jest/__mock__';
@@ -58,6 +59,16 @@ const renderMARCTableViewRow = ({
 };
 
 describe('MARCTableViewRow', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderMARCTableViewRow(MARCTableViewRowProps({
+      action: 'EDIT',
+      subaction: 'CREATE_NEW_FIELD',
+      text: 'testText',
+    }));
+
+    await runAxeTest({ rootNode: container });
+  });
+
   describe('when mapping details action is Edit', () => {
     describe('when mapping details subaction is not Replace', () => {
       describe('when row has the text ', () => {