+++ src/components/MatchCriterion/edit/MatchCriterion.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../../test/jest/__mock__';
@@ -105,6 +106,12 @@ const renderMatchCriterion = ({
 };
 
 describe('MatchCriterion edit', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderMatchCriterion({ ...matchCriterionProps('INSTANCE', 'MARC_BIBLIOGRAPHIC', 'BEGINS_WITH') });
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should be rendered with a qualifier type', () => {
     const { getByText } = renderMatchCriterion({ ...matchCriterionProps('INSTANCE', 'MARC_BIBLIOGRAPHIC', 'BEGINS_WITH') });