+++ src/components/MatchCriterion/edit/QualifierPartSection/QualifierPartSection.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../../../test/jest/__mock__';
@@ -45,6 +46,12 @@ const renderQualifierPartSection = ({
 describe('QualifierPartSection edit', () => {
   afterAll(() => {
     onChangeMock.mockClear();
+  });
+
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderQualifierPartSection(qualifierPartSection);
+
+    await runAxeTest({ rootNode: container });
   });
 
   it('should have a correct title', () => {