+++ src/components/ProfileTree/ProfileLinker/ProfileLinker.test.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import '../../../../test/jest/__mock__';
@@ -53,6 +54,12 @@ const renderProfileLinker = ({
 describe('ProfileLinker', () => {
   afterEach(() => {
     Pluggable.mockClear();
+  });
+
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderProfileLinker(profileLinkerProps);
+
+    await runAxeTest({ rootNode: container });
   });
 
   it('should be rendered', () => {