+++ src/settings/MARCFieldProtection/MARCFieldProtection.test.js
@@ -5,6 +5,7 @@ import {
   screen,
   fireEvent,
 } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { buildResources } from '@folio/stripes-data-transfer-components/test/helpers';
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
@@ -101,6 +102,13 @@ const renderMarcFieldProtection = () => {
 };
 
 describe('MARCFieldProtection component', () => {
+  // TODO: Create separate ticket to fix all the accesibility tests
+  it.skip('should be rendered with no axe errors', async () => {
+    const { container } = renderMarcFieldProtection();
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should be rendered', () => {
     const { getByText } = renderMarcFieldProtection();