+++ src/settings/MappingProfiles/detailsSections/edit/InstanceDetailsSections/tests/ElectronicAccess.test.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 
@@ -64,6 +65,12 @@ const renderElectronicAccess = () => {
 };
 
 describe('<ElectronicAccess>', () => {
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderElectronicAccess();
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should render correct fields', async () => {
     const { getByText } = renderElectronicAccess();