+++ src/settings/MappingProfiles/detailsSections/edit/InvoiceDetailSection/VendorInformation.js
@@ -2,6 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
 import { Field } from 'redux-form';
+import { isEmpty } from 'lodash';
 
 import {
   Accordion,
@@ -15,7 +16,10 @@ import {
   AcceptedValuesField,
 } from '../../../../../components';
 
-import { getFieldName } from '../../utils';
+import {
+  getFieldEnabled,
+  getFieldName,
+} from '../../utils';
 import {
   okapiShape,
   validateRequiredField,
@@ -27,8 +31,19 @@ export const VendorInformation = ({
   filledVendorId,
   accountingCodeOptions,
   onSelectVendor,
+  onClearVendor,
   okapi,
 }) => {
+  const VENDOR_INFO_FIELDS_MAP = {
+    VENDOR_INVOICE_NUMBER: getFieldName(16),
+    VENDOR_NAME: getFieldName(17),
+    ACCOUNTING_CODE: 18,
+  };
+
+  const onAccountingCodeChange = value => {
+    setReferenceTables(getFieldEnabled(VENDOR_INFO_FIELDS_MAP.ACCOUNTING_CODE), !isEmpty(value));
+  };
+
   return (
     <Accordion
       id="vendor-information"
@@ -39,7 +54,7 @@ export const VendorInformation = ({
           <Field
             component={TextField}
             label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.vendorInformation.field.vendorInvoiceNumber`} />}
-            name={getFieldName(16)}
+            name={VENDOR_INFO_FIELDS_MAP.VENDOR_INVOICE_NUMBER}
             validate={validateRequiredField}
             required
           />
@@ -48,8 +63,9 @@ export const VendorInformation = ({
           <FieldOrganization
             id={filledVendorId}
             setReferenceTables={setReferenceTables}
-            name={getFieldName(17)}
+            name={VENDOR_INFO_FIELDS_MAP.VENDOR_NAME}
             onSelect={onSelectVendor}
+            onClear={onClearVendor}
             label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.vendorInformation.field.vendorName`} />}
             validate={validateRequiredField}
             required
@@ -59,13 +75,14 @@ export const VendorInformation = ({
         <Col xs={4}>
           <AcceptedValuesField
             component={TextField}
-            name={getFieldName(18)}
+            name={getFieldName(VENDOR_INFO_FIELDS_MAP.ACCOUNTING_CODE)}
             label={<FormattedMessage id={`${TRANSLATION_ID_PREFIX}.invoice.vendorInformation.field.accountingCode`} />}
             optionValue="value"
             optionLabel="label"
             wrapperLabel={`${TRANSLATION_ID_PREFIX}.wrapper.acceptedValues`}
             acceptedValuesList={accountingCodeOptions}
             disabled={!accountingCodeOptions.length}
+            onChange={onAccountingCodeChange}
             okapi={okapi}
           />
         </Col>
@@ -78,6 +95,7 @@ VendorInformation.propTypes = {
   okapi: okapiShape.isRequired,
   setReferenceTables: PropTypes.func.isRequired,
   onSelectVendor: PropTypes.func.isRequired,
+  onClearVendor: PropTypes.func,
   filledVendorId: PropTypes.string,
   accountingCodeOptions: PropTypes.arrayOf(PropTypes.object),
 };