+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/FundDistribution.test.js
@@ -3,6 +3,7 @@ import {
   fireEvent,
   within,
 } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 
@@ -72,6 +73,12 @@ const renderFundDistribution = () => {
 describe('FundDistribution', () => {
   afterEach(() => {
     setReferenceTablesMock.mockClear();
+  });
+
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderFundDistribution();
+
+    await runAxeTest({ rootNode: container });
   });
 
   it('should render correct fields', async () => {