+++ src/settings/MappingProfiles/detailsSections/edit/OrderDetailSection/tests/OrderInformation.test.js
@@ -3,6 +3,7 @@ import {
   fireEvent,
   within,
 } from '@testing-library/react';
+import { runAxeTest } from '@folio/stripes-testing';
 
 import { renderWithIntl } from '@folio/stripes-data-transfer-components/test/jest/helpers';
 import { buildResources } from '@folio/stripes-data-transfer-components/test/helpers';
@@ -101,6 +102,12 @@ describe('OrderInformation', () => {
     setReferenceTablesMock.mockClear();
   });
 
+  it('should be rendered with no axe errors', async () => {
+    const { container } = renderOrderInformation();
+
+    await runAxeTest({ rootNode: container });
+  });
+
   it('should render correct fields', async () => {
     const { getByText } = renderOrderInformation();
 
@@ -128,6 +135,7 @@ describe('OrderInformation', () => {
   it('should render required fields', () => {
     const { queryByText } = renderOrderInformation();
 
+    expect(within(queryByText('Purchase order status')).getByText(/\*/i)).toBeDefined();
     expect(within(queryByText('Vendor')).getByText(/\*/i)).toBeDefined();
     expect(within(queryByText('Order type')).getByText(/\*/i)).toBeDefined();
   });